+++ flask_jsonapi/resources.py
@@ -1,7 +1,6 @@
 import http
 import logging
 
-import flask
 import marshmallow
 from flask import helpers
 from flask import request
@@ -26,6 +25,8 @@ class ResourceBase(views.View):
     def __init__(self, *, schema=None):
         if schema:
             self.schema = schema
+        self.include_parser = query_string.IncludeParser(schema=self.schema)
+        self.sparse_fields_parser = query_string.SparseFieldsParser(schema=self.schema)
 
     @classmethod
     def as_view(cls, name, *class_args, **class_kwargs):
@@ -53,18 +54,6 @@ class ResourceBase(views.View):
         )
         return helpers.make_response('Method Not Allowed.', http.HTTPStatus.METHOD_NOT_ALLOWED)
 
-    def _check_include_fields(self):
-        include_parameter = flask.request.args.get('include')
-        if include_parameter:
-            include_fields = tuple(include_parameter.split(','))
-            try:
-                self.schema().check_relations(include_fields)
-            except ValueError as exc:
-                raise exceptions.InvalidInclude(detail=str(exc))
-            return include_fields
-        else:
-            return tuple()
-
 
 class ResourceDetail(ResourceBase):
     methods = ['GET', 'DELETE', 'PATCH']
@@ -72,9 +61,10 @@ class ResourceDetail(ResourceBase):
 
     def get(self, *args, **kwargs):
         resource = self.read(self.resource_id)
-        include_fields = self._check_include_fields()
+        include_fields = self.include_parser.parse()
+        sparse_fields = self.sparse_fields_parser.parse()
         try:
-            data, errors = self.schema(include_data=include_fields).dump(resource)
+            data, errors = self.schema(include_data=include_fields, only=sparse_fields).dump(resource)
         except marshmallow.ValidationError as e:
             return response.JsonApiErrorResponse.from_marshmallow_errors(e.messages)
         else:
@@ -132,11 +122,19 @@ class ResourceList(ResourceBase):
         parsed_pagination = self.pagination.parse()
         objects_list = self.read_many(filters=parsed_filters,
                                       pagination=parsed_pagination)
-        include_fields = self._check_include_fields()
+        include_fields = self.include_parser.parse()
+        sparse_fields = self.sparse_fields_parser.parse()
         try:
-            objects, errors = self.schema(many=True, include_data=include_fields).dump(objects_list)
+            objects, errors = self.schema(
+                many=True, include_data=include_fields, only=sparse_fields).dump(objects_list)
         except marshmallow.ValidationError as e:
             return response.JsonApiErrorResponse.from_marshmallow_errors(e.messages)
+        except (AttributeError, KeyError, ValueError) as e:
+            logger.error(
+                'Error Processing Request',
+                extra={'status_code': http.HTTPStatus.BAD_REQUEST, 'request': request, 'exception': e}
+            )
+            return helpers.make_response('Error Processing Request', http.HTTPStatus.BAD_REQUEST)
         else:
             if errors:
                 return response.JsonApiErrorResponse.from_marshmallow_errors(errors)